Factorized engineHelpers.
With a small repository it is nicer to have each helper in a separate file.
